+++ src/main/java/com/esprit/achat/services/Interface/AdresseService.java
@@ -2,5 +2,5 @@ package com.esprit.achat.services.Interface;
 
 import com.esprit.achat.persistence.entity.Adresse;
 
-public interface AdresseService extends CrudService<Adresse,Integer> {
+public interface AdresseService extends CrudService<Adresse,Long> {
 }
